moved train functions to seperate files, custom parameter support instead of scale/bias, updated templates
