+++ experiments/Web/webNN_template_GA.py
@@ -5,9 +5,8 @@ Created on Thu Nov 29 18:07:01 2018
 
 @author: ljknoll
 
-testing evolution script for training web of neural networks
+Testing evolution script for training web of neural networks, both GA and GD
 
-removed dependency on config.config_class
 """
 import torch
 import matplotlib.pyplot as plt
@@ -29,13 +28,11 @@ net = predNNet(nn_file)
 # build web
 web = webNNet()
 web.add_vertex(net, 'A', output=True)
-#web.add_vertex(net, 'B')
-#web.add_arc('B', 'A', 3)
 
 
 # batch size
 N = 7
-# input data, size (N, 2*number of networks)
+# input data, size (N, 2 * nr of networks)
 x = torch.cat((torch.zeros(N), torch.linspace(-0.9, 0.9, N))).view(2, -1).t()
 # target data, size (N, nr output vertices)
 target = torch.linspace(0, 0.5, N).view(-1,1)
@@ -57,9 +54,8 @@ plt.title('GA output')
 
 
 web.reset_parameters()
-
-
 loss, params = web.train(x, target, 1, 200, lr=0.05, beta=0.1)
+
 web.reset_parameters(params)
 web.forward(x)
